refactor: move write_floxhub_netrc
Since this function is only going to be used from the auth provider, it
makes sense to move this here. I'm hoping that we can move _all_ auth
logic to this provider over time.
