+++ cli/flox/src/commands/upload.rs
@@ -3,7 +3,8 @@ use std::path::PathBuf;
 use anyhow::{Context, Result, bail};
 use bpaf::Bpaf;
 use flox_rust_sdk::flox::Flox;
-use flox_rust_sdk::providers::publish::{ClientSideCatalogStoreConfig, write_floxhub_netrc};
+use flox_rust_sdk::providers::auth::write_floxhub_netrc;
+use flox_rust_sdk::providers::publish::ClientSideCatalogStoreConfig;
 use tracing::instrument;
 use url::Url;
